Added a benchmark project. Currently only one benchmark which compares performance to wasmtime. This project actually beats wasmtime!
The benchmark runs a function called `bench`, which is adapted from a random Rust project I had lying around (calculating various rocket nozzle parameters). I'm happy to share the code for this if you want, but I'm not sure what the best way to do it is - would you just want the entire Rust project included in the wasm folder?